Issue-201: Moved keywordsview package to separate module.
