+++ core/src/main/res/values/dimens.xml
@@ -32,12 +32,6 @@
     <dimen name="trailer_item_margin">16dp</dimen>
     <dimen name="trailer_item_source_image_size">48dp</dimen>
 
-    <dimen name="keyword_item_corner_radius">16dp</dimen>
-    <dimen name="keyword_item_padding">4dp</dimen>
-    <dimen name="keyword_item_top_margin">2dp</dimen>
-    <dimen name="keyword_item_bottom_margin">2dp</dimen>
-    <dimen name="keyword_item_edges_margin">4dp</dimen>
-
     <dimen name="credit_item_height">250dp</dimen>
     <dimen name="credit_item_width">150dp</dimen>
     <dimen name="credit_item_image_height">190dp</dimen>
