+++ keywordsview/build.gradle
@@ -1,0 +1,7 @@
+apply from: '../buildsystem/android-module.gradle'
+
+dependencies {
+    implementation project(path: ':core')
+
+    implementation libraries.flexbox
+}
