+++ keywordsview/src/main/java/com/kirchhoff/movies/keywordsview/adapter/KeywordVH.kt
@@ -1,10 +1,10 @@
-package com.kirchhoff.movies.ui.views.keywords.adapter
+package com.kirchhoff.movies.keywordsview.adapter
 
 import android.view.ViewGroup
 import android.widget.TextView
-import com.kirchhoff.movies.R
 import com.kirchhoff.movies.core.extensions.inflate
 import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
+import com.kirchhoff.movies.keywordsview.R
 
 class KeywordVH(parent: ViewGroup) : BaseVH<String>(parent.inflate(R.layout.item_keyword)) {
