+++ keywordsview/src/main/res/values/dimens.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="keyword_item_corner_radius">16dp</dimen>
+    <dimen name="keyword_item_padding">4dp</dimen>
+    <dimen name="keyword_item_top_margin">2dp</dimen>
+    <dimen name="keyword_item_bottom_margin">2dp</dimen>
+    <dimen name="keyword_item_edges_margin">4dp</dimen>
+</resources>
